+++ eslint.config.mjs
@@ -6,12 +6,13 @@ export default antfu(
     react: true,
     typescript: true,
     // apps are independent lints
-    ignores: ['apps/**'],
+    // ignores: ['apps/**'],
   },
   {
     languageOptions: { globals: globals.browser },
     rules: {
       'no-console': 'off',
+      'node/prefer-global/process': 'off',
       'react/no-forward-ref': 'off',
       'react/no-use-context': 'off',
       '@typescript-eslint/no-empty-object-type': 'off',