+++ packages/ui/button/src/button.tsx
@@ -1,4 +1,3 @@
-import type { ButtonProps as NativeButtonProps } from '@tarojs/components'
 import type { UseButtonProps } from './use'
 import { Spinner } from '@srcube-taro/spinner'
 import { forwardRef } from 'react'
@@ -6,7 +5,7 @@ import { useButton } from './use'
 
 export interface ButtonProps extends UseButtonProps {}
 
-const Button = forwardRef<NativeButtonProps, ButtonProps>((props, ref) => {
+const Button = forwardRef<any, ButtonProps>((props, ref) => {
   const {
     Component,
     domRef,