+++ packages/ui/spinner/src/spinner.tsx
@@ -1,4 +1,3 @@
-import type { ViewProps } from '@tarojs/components'
 import type { UseSpinnerProps } from './use'
 import { View } from '@tarojs/components'
 import { forwardRef } from 'react'
@@ -6,7 +5,7 @@ import { useSpinner } from './use'
 
 export interface SpinnerProps extends UseSpinnerProps { }
 
-const Spinner = forwardRef<ViewProps, SpinnerProps>((props, ref) => {
+const Spinner = forwardRef<any, SpinnerProps>((props, ref) => {
   const { Component, domRef, label, getSpinnerProps, styles } = useSpinner({ ...props, ref })
 
   return (