+++ packages/utils/react/src/index.ts
@@ -1,2 +1,4 @@
 export * from './create-context'
+export * from './dom'
+export * from './state'
 export * from './types'