+++ packages/core/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @srcube-taro/theme
 
+## 0.0.2
+
+### Patch Changes
+
+- Code styles and components ref type
+
 ## 0.0.1
 
 ### Patch Changes