+++ packages/ui/layouts/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @srcube-taro/layouts
 
+## 0.0.2
+
+### Patch Changes
+
+- Code styles and components ref type
+- Updated dependencies
+  - @srcube-taro/utils-react@0.0.3
+  - @srcube-taro/theme@0.0.2
+  - @srcube-taro/utils-taro@0.0.3
+  - @srcube-taro/utils-tv@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes